First Commit
I have designed a small terrain area that is complete with a starting environment, skybox and a playable character. There are also functioning enemy NPC's in the game and a starting main menu that has a few buttons working. I am now working on making sure that the player can shoot, pick up health and setting up a camera switch fucntion.