Made a number of changes
When i first loaded up the file i found that the game was essentially broken but it wasnt till later that i found out it was because of the SHooting Script that i had. I deleted it and fixed the game and from then i have added a number of changes to the game such as removing th tress from the terrain, adjusting certain scripts and so forth.